Multiple changes + Time is now measured in fractions of a second, not microseconds + Viewports in RGSS1 will now only retain weak references to their children + Invalid Sprite bitmaps found during Graphics.update will be treated as null + Bitmap#mega? and Bitmap#animated? are properly rescued

+++ src/display/graphics.h
@@ -36,8 +36,8 @@ struct Movie;
 class Graphics
 {
 public:
-    unsigned long long getDelta();
-    unsigned long long lastUpdate();
+    double getDelta();
+    double lastUpdate();
     
 	void update(bool checkForShutdown = true);
 	void freeze();
