+++ src/input/input.h
@@ -58,7 +58,7 @@ public:
     
     void recalcRepeat(unsigned int fps);
 
-    unsigned long long getDelta();
+    double getDelta();
 	void update();
     
     std::vector<std::string> getBindings(ButtonCode code);
@@ -68,21 +68,21 @@ public:
 	bool isRepeated(int button);
     bool isReleased(int button);
     unsigned int count(int button);
-    unsigned long long repeatTime(int button);
+    double repeatTime(int button);
     
     bool isPressedEx(int code, bool isVKey);
     bool isTriggeredEx(int code, bool isVKey);
     bool isRepeatedEx(int code, bool isVKey);
     bool isReleasedEx(int code, bool isVKey);
     unsigned int repeatcount(int code, bool isVKey);
-    unsigned long long repeatTimeEx(int code, bool isVKey);
+    double repeatTimeEx(int code, bool isVKey);
     
     bool controllerIsPressedEx(int button);
     bool controllerIsTriggeredEx(int button);
     bool controllerIsRepeatedEx(int button);
     bool controllerIsReleasedEx(int button);
     unsigned int controllerRepeatcount(int button);
-    unsigned long long controllerRepeatTimeEx(int button);
+    double controllerRepeatTimeEx(int button);
     
     uint8_t *rawKeyStates();
     unsigned int rawKeyStatesLength();
