+++ src/sharedstate.h
@@ -86,7 +86,7 @@ struct SharedState
 	unsigned int genTimeStamp();
     
     // Returns time since SharedState was constructed in microseconds
-    unsigned long long runTime();
+    double runTime();
 
 	/* Returns global quad IBO, and ensures it has indices
 	 * for at least minSize quads */
